refactor(slugPrefix): move slugPrefix inside generateHierarchicalSlug.
- Add index for utils for easier import.
- Move slugPrefix inside generateHierarchicalSlug.
- Remove slugPrefix from TOCLink.
