+++ src/services/utils/index.ts
@@ -1,0 +1,5 @@
+export { default as getFilteredPosts } from './getFilteredPosts';
+export {
+  default as generateHierarchicalSlug,
+  slugPrefix,
+} from './generateHierarchicalSlug';
